Restructure polls directory
