+++ src/commands/irregulars.ts
@@ -23,7 +23,7 @@ import {
   isMemberInRegulars,
   isMemberInVip,
 } from '../utils/members.js';
-import { createPoll, isPollDuplicate } from '../utils/polls/main.js';
+import { createPoll, isPollDuplicate } from '../utils/polls/core/special.js';
 
 const name = 'irregulars';
