+++ src/commands/vip.ts
@@ -24,7 +24,7 @@ import {
   isMemberInIrregulars,
   isMemberInVip,
 } from '../utils/members.js';
-import { createPoll, isPollDuplicate } from '../utils/polls/main.js';
+import { createPoll, isPollDuplicate } from '../utils/polls/core/special.js';
 
 const name = 'vip';
