+++ src/events/ClientReady.ts
@@ -6,7 +6,7 @@ import { logger } from '../logger.js';
 import { bootMessage, logMessageFunctions } from '../translations/logs.js';
 import { getChannel, initializeChannels } from '../utils/channels.js';
 import { initializeCronJobs } from '../utils/cron/main.js';
-import { initializePolls } from '../utils/polls/main.js';
+import { initializePolls } from '../utils/polls/core/special.js';
 import { initializeRoles } from '../utils/roles.js';
 
 export const name = Events.ClientReady;
